Streamlit/requirements.txt
nltk==3.7
numpy==1.23.1
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.3.0
requests==2.28.1
streamlit==1.11.1
tensorflow==2.9.1
